refactor: axios types 모듈 분리
