+++ src/axios/event/getEvent.ts
@@ -1,4 +1,5 @@
-import request, { ResType } from '@axios/index';
+import request from '@axios/index';
+import { ResType } from '@axios/types';
 
 interface GetEventParamTypes {
   eventId: string | undefined | string[];
