--- src/contexts/types.ts
@@ -1,4 +1,0 @@
-export interface ErrorType {
-  code: number | null;
-  message: string | null;
-}
